COLLECTOR-245. Better handle SIGTERM in sdc
This requires a corresponding patch in the datacollector repo.

More details of the investigation can be found in ESC-1290, but the gist is that
we don't properly propagate the SIGINT signal down to the JVM, so SDC is not
gracefully shutdown (i.e. skips the shutdown hook).

This patch changes the docker-entrypoint.sh logic so that it no longer sits
around simply to propagate the signal to the underlying process (i.e. _sdc).  It
wasn't handling some cases, and just isn't necessary to have this extra layer.
Instead, it simply execs, making _sdc ultimately PID 1; and with the
corresponding patch in the datacollector repo, _sdc knows how to handle the
signals now.

Manually tested that the docker container behaves correctly with and without
being registered to control hub.  The shutdown hook is called when doing ctrl-c,
docker stop (SIGTERM), shutdown/restart in SCH, etc.

Change-Id: Ifffee10ebcec7dec65694397391e9a776ddfd6ea