COLLECTOR-5006. Support for Custom CA certs
When present in the base image, as is the case with the eclipse-temurin
family, we'll preserve the custom CA cert entrypoint from upstream.

If a cert is provided via the $CUSTOM_TRUSTSTORE_CA_CERT environment
variable, we'll explicitly add it.  This allows us to provide it via
an "install command" from SCH.

Change-Id: I90a77db8feba9b345cc007b8be5cbe2899b9691c